requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
tweepy==4.14.0
urllib3==2.2.3
Werkzeug==3.1.2
